requirement.txt
python-telegram-bot >= 13.11
colorama
pandas
pyyaml